feat(architecture): move data models to the separate folder in persistence layer.
