+++ src/FlashCards.Infrastructure/Persistence/DataModels/FlashCardsUnit.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace FlashCards.Core.Domain.Entities;
+namespace FlashCards.Infrastructure.Persistence.DataModels;
 
 public class FlashCardsUnit
 {
