+++ src/FlashCards.Infrastructure/Persistence/DataModels/User.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using FlashCards.Core.Domain.Constants;
 
-namespace FlashCards.Core.Domain.Entities;
+namespace FlashCards.Infrastructure.Persistence.DataModels;
 
 public class User
 {
