+++ src/FlashCards.Infrastructure/Persistence/Repositories/UserRepository.cs
@@ -1,5 +1,6 @@
 ﻿using FlashCards.Core.Domain.Entities;
 using FlashCards.Infrastructure.Persistence.Contexts;
+using FlashCards.Infrastructure.Persistence.DataModels;
 using Microsoft.EntityFrameworkCore;
 
 namespace FlashCards.Infrastructure.Persistence.Repositories;
